requirements.txt
openai>=1.0.0
openai-whisper
moviepy==1.0.3
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
edge-tts>=6.1.0
schedule>=1.2.0
pytz>=2023.3
langgraph>=0.0.20
langchain-openai>=0.0.5
opencv-python-headless>=4.8.0
gtts>=2.5.0
duckduckgo_search>=4.0.0
langchain-core>=0.1.0
imageio>=2.34.0
